+++ Player.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "Level.h"
+#include "Economy.h"
 #include <SFML/Graphics.hpp>
 
 class Player
@@ -12,8 +13,11 @@ class Player
     void move(float x, float y);
     void setStartPosition();
 
+    Economy& getEconomy(); // Accessor for player economy
+
   private:
     sf::CircleShape shape;
     sf::Vector2f position;
     const Level &level;
+    Economy economy;
 };